+++ src/wiktextract/extractor/es/translation.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from wikitextprocessor import WikiNode
 
-from wiktextract.extractor.de.utils import split_senseids
+from wiktextract.extractor.share import split_senseids
 from wiktextract.extractor.es.models import Translation, WordEntry
 from wiktextract.page import clean_node
 from wiktextract.wxr_context import WiktextractContext
