+++ src/wiktextract/extractor/share.py
@@ -90,3 +90,31 @@ def split_tag_text(text: str) -> List[str]:
         tag.strip("()").strip()
         for tag in re.split(r"(?<=\))\s+(?=\()", text.strip())
     ]
+
+
+def split_senseids(senseids_str: str) -> list[str]:
+    senseids = []
+    raw_ids = (
+        senseids_str.strip().removeprefix("[").removesuffix("]").split(",")
+    )
+    for raw_id in raw_ids:
+        range_split = raw_id.split("-")
+        if len(range_split) == 1:
+            senseids.append(raw_id.strip())
+        elif len(range_split) == 2:
+            try:
+                start = re.sub(r"[a-z]", "", range_split[0].strip())
+                end = re.sub(r"[a-z]", "", range_split[1].strip())
+                senseids.extend(
+                    [
+                        str(id)
+                        for id in range(
+                            int(start),
+                            int(end) + 1,
+                        )
+                    ]
+                )
+            except:
+                pass
+
+    return senseids
